refactor: move drawing logic to its own component
