+++ examples/ai-lab-example/package.json
@@ -4,7 +4,7 @@
   "version": "1.0.0",
   "license": "MIT",
   "scripts": {
-    "start": "cd../../ && pnpm build --filter ai-lab && pnpm build --filter @ai-lab/web-example && cd examples/ai-lab-example && pnpm storybook",
+    "start": "cd ../../ && pnpm build --filter ai-lab && pnpm build --filter @ai-lab/web-example && cd examples/ai-lab-example && pnpm storybook",
     "build": "parcel build index.html",
     "storybook": "start-storybook -p 6006 -c storybook",
     "build-storybook": "build-storybook"
