+++ examples/ai-lab-example/tsconfig.json
@@ -4,5 +4,5 @@
     "rootDir": "."
   },
   "extends": "../../shared/tsconfig.base.json",
-  "include": ["."]
+  "include": [".", "../../packages/ai-lab/src"]
 }
