+++ package.json
@@ -67,6 +67,7 @@
     "@types/react": "^17.0.22",
     "@types/react-dom": "^17.0.9",
     "husky": "^7.0.2",
+    "prettier": "^2.5.1",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "size-limit": "^5.0.3",
