+++ packages/ai-lab/src/components/AILabVideo/index.ts
@@ -1,3 +1,2 @@
 export * from './AILabWebCam';
 export * from './AILabLocalVideo';
-export * from './types';
