--- packages/ai-lab/src/components/AILabVideo/types.ts
@@ -1,8 +1,0 @@
-import { PerformanceProps } from '../../performance';
-
-export interface VideoProps
-  extends React.DetailedHTMLProps<
-      React.VideoHTMLAttributes<HTMLVideoElement>,
-      HTMLVideoElement
-    >,
-    PerformanceProps {}
